Move autocomplete list generation to RouterList
