+++ src/components/Search/SearchRouter/SearchRouter.tsx
@@ -1,35 +1,23 @@
 import {useNavigationState} from '@react-navigation/native';
-import {Str} from 'expensify-common';
-import isEmpty from 'lodash/isEmpty';
 import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 import {View} from 'react-native';
 import type {TextInputProps} from 'react-native';
 import {useOnyx} from 'react-native-onyx';
-import type {ValueOf} from 'type-fest';
 import HeaderWithBackButton from '@components/HeaderWithBackButton';
+import * as Expensicons from '@components/Icon/Expensicons';
 import {usePersonalDetails} from '@components/OnyxProvider';
 import {useOptionsList} from '@components/OptionListContextProvider';
-import type {SearchAutocompleteQueryRange, SearchQueryString} from '@components/Search/types';
+import type {SearchQueryString} from '@components/Search/types';
 import type {SelectionListHandle} from '@components/SelectionList/types';
-import useActiveWorkspace from '@hooks/useActiveWorkspace';
 import useDebouncedState from '@hooks/useDebouncedState';
 import useKeyboardShortcut from '@hooks/useKeyboardShortcut';
 import useLocalize from '@hooks/useLocalize';
-import usePolicy from '@hooks/usePolicy';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
-import * as CardUtils from '@libs/CardUtils';
 import * as OptionsListUtils from '@libs/OptionsListUtils';
 import {getAllTaxRates} from '@libs/PolicyUtils';
 import type {OptionData} from '@libs/ReportUtils';
-import {
-    getAutocompleteCategories,
-    getAutocompleteRecentCategories,
-    getAutocompleteRecentTags,
-    getAutocompleteTags,
-    getAutocompleteTaxList,
-    parseForAutocomplete,
-} from '@libs/SearchAutocompleteUtils';
+import {parseForAutocomplete} from '@libs/SearchAutocompleteUtils';
 import * as SearchQueryUtils from '@libs/SearchQueryUtils';
 import Navigation from '@navigation/Navigation';
 import variables from '@styles/variables';
@@ -38,13 +26,11 @@ import Timing from '@userActions/Timing';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
-import type PersonalDetails from '@src/types/onyx/PersonalDetails';
 import {getQueryWithSubstitutions} from './getQueryWithSubstitutions';
 import type {SubstitutionMap} from './getQueryWithSubstitutions';
 import {getUpdatedSubstitutionsMap} from './getUpdatedSubstitutionsMap';
 import SearchRouterInput from './SearchRouterInput';
 import SearchRouterList from './SearchRouterList';
-import type {AutocompleteItemData} from './SearchRouterList';
 
 type SearchRouterProps = {
     onRouterClose: () => void;
@@ -57,13 +43,21 @@ function SearchRouter({onRouterClose, shouldHideInputCaret}: SearchRouterProps) 
     const [betas] = useOnyx(ONYXKEYS.BETAS);
     const [recentSearches] = useOnyx(ONYXKEYS.RECENT_SEARCHES);
     const [isSearchingForReports] = useOnyx(ONYXKEYS.IS_SEARCHING_FOR_REPORTS, {initWithStoredValues: false});
-    const [autocompleteSuggestions, setAutocompleteSuggestions] = useState<AutocompleteItemData[] | undefined>([]);
     const [autocompleteSubstitutions, setAutocompleteSubstitutions] = useState<SubstitutionMap>({});
+
+    const personalDetails = usePersonalDetails();
+    const [reports = {}] = useOnyx(ONYXKEYS.COLLECTION.REPORT);
+    const [cardList = {}] = useOnyx(ONYXKEYS.CARD_LIST);
+    const taxRates = getAllTaxRates();
 
     const {shouldUseNarrowLayout} = useResponsiveLayout();
     const listRef = useRef<SelectionListHandle>(null);
 
+    // The actual input text that the user sees
     const [textInputValue, debouncedInputValue, setTextInputValue] = useDebouncedState('', 500);
+    // The input text that was last used for autocomplete; needed for the SearchRouterList when browsing list via arrow keys
+    const [autocompleteInputValue, setAutocompleteInputValue] = useState(textInputValue);
+
     const contextualReportID = useNavigationState<Record<string, {reportID: string}>, string | undefined>((state) => {
         return state?.routes.at(-1)?.params?.reportID;
     });
@@ -105,253 +99,89 @@ function SearchRouter({onRouterClose, shouldHideInputCaret}: SearchRouterProps) 
             return searchOptions.recentReports.slice(0, 10);
         }
 
-        const reports: OptionData[] = [...filteredOptions.recentReports, ...filteredOptions.personalDetails];
+        const reportOptions: OptionData[] = [...filteredOptions.recentReports, ...filteredOptions.personalDetails];
         if (filteredOptions.userToInvite) {
-            reports.push(filteredOptions.userToInvite);
+            reportOptions.push(filteredOptions.userToInvite);
         }
-        return reports.slice(0, 10);
+        return reportOptions.slice(0, 10);
     }, [debouncedInputValue, filteredOptions, searchOptions]);
 
     const contextualReportData = contextualReportID ? searchOptions.recentReports?.find((option) => option.reportID === contextualReportID) : undefined;
 
-    const {activeWorkspaceID} = useActiveWorkspace();
-    const policy = usePolicy(activeWorkspaceID);
-
-    const typeAutocompleteList = Object.values(CONST.SEARCH.DATA_TYPES);
-    const statusAutocompleteList = Object.values({...CONST.SEARCH.STATUS.TRIP, ...CONST.SEARCH.STATUS.INVOICE, ...CONST.SEARCH.STATUS.CHAT, ...CONST.SEARCH.STATUS.TRIP});
-    const expenseTypes = Object.values(CONST.SEARCH.TRANSACTION_TYPE);
-    const [cardList = {}] = useOnyx(ONYXKEYS.CARD_LIST);
-    const cardAutocompleteList = Object.values(cardList);
-    const personalDetailsForParticipants = usePersonalDetails();
-
-    const participantsAutocompleteList = useMemo(
-        () =>
-            Object.values(personalDetailsForParticipants)
-                .filter((details): details is NonNullable<PersonalDetails> => !!(details && details?.login))
-                .map((details) => ({
-                    name: details.displayName ?? Str.removeSMSDomain(details.login ?? ''),
-                    accountID: details?.accountID.toString(),
-                })),
-        [personalDetailsForParticipants],
-    );
-    const allTaxRates = getAllTaxRates();
-    const taxAutocompleteList = useMemo(() => getAutocompleteTaxList(allTaxRates, policy), [policy, allTaxRates]);
-    const [allPolicyCategories] = useOnyx(ONYXKEYS.COLLECTION.POLICY_CATEGORIES);
-    const [allRecentCategories] = useOnyx(ONYXKEYS.COLLECTION.POLICY_RECENTLY_USED_CATEGORIES);
-    const categoryAutocompleteList = useMemo(() => {
-        return getAutocompleteCategories(allPolicyCategories, activeWorkspaceID);
-    }, [activeWorkspaceID, allPolicyCategories]);
-    const recentCategoriesAutocompleteList = useMemo(() => {
-        return getAutocompleteRecentCategories(allRecentCategories, activeWorkspaceID);
-    }, [activeWorkspaceID, allRecentCategories]);
-
-    const [currencyList] = useOnyx(ONYXKEYS.CURRENCY_LIST);
-    const currencyAutocompleteList = Object.keys(currencyList ?? {});
-    const [recentCurrencyAutocompleteList] = useOnyx(ONYXKEYS.RECENTLY_USED_CURRENCIES);
-
-    const [allPoliciesTags] = useOnyx(ONYXKEYS.COLLECTION.POLICY_TAGS);
-    const [allRecentTags] = useOnyx(ONYXKEYS.COLLECTION.POLICY_RECENTLY_USED_TAGS);
-    const tagAutocompleteList = useMemo(() => {
-        return getAutocompleteTags(allPoliciesTags, activeWorkspaceID);
-    }, [activeWorkspaceID, allPoliciesTags]);
-    const recentTagsAutocompleteList = getAutocompleteRecentTags(allRecentTags, activeWorkspaceID);
-
-    const updateAutocomplete = useCallback(
-        (autocompleteValue: string, ranges: SearchAutocompleteQueryRange[], autocompleteType?: ValueOf<typeof CONST.SEARCH.SYNTAX_ROOT_KEYS & typeof CONST.SEARCH.SYNTAX_FILTER_KEYS>) => {
-            const alreadyAutocompletedKeys: string[] = [];
-            ranges.forEach((range) => {
-                if (!autocompleteType || range.key !== autocompleteType) {
-                    return;
-                }
-                alreadyAutocompletedKeys.push(range.value.toLowerCase());
-            });
-
-            let filteredAutocompleteSuggestions: AutocompleteItemData[] | undefined;
-            switch (autocompleteType) {
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.TAG: {
-                    const autocompleteList = autocompleteValue ? tagAutocompleteList : recentTagsAutocompleteList ?? [];
-                    const filteredTags = autocompleteList
-                        .filter((tag) => tag.toLowerCase()?.includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(tag))
-                        .sort()
-                        .slice(0, 10);
-
-                    filteredAutocompleteSuggestions = filteredTags.map((tagName) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.TAG,
-                        text: tagName,
-                    }));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.CATEGORY: {
-                    const autocompleteList = autocompleteValue ? categoryAutocompleteList : recentCategoriesAutocompleteList;
-                    const filteredCategories = autocompleteList
-                        .filter((category) => {
-                            return category.toLowerCase()?.includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(category.toLowerCase());
-                        })
-                        .sort()
-                        .slice(0, 10);
-
-                    filteredAutocompleteSuggestions = filteredCategories.map((categoryName) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.CATEGORY,
-                        text: categoryName,
-                    }));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.CURRENCY: {
-                    const autocompleteList = autocompleteValue ? currencyAutocompleteList : recentCurrencyAutocompleteList ?? [];
-                    const filteredCurrencies = autocompleteList
-                        .filter((currency) => currency.toLowerCase()?.includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(currency.toLowerCase()))
-                        .sort()
-                        .slice(0, 10);
-
-                    filteredAutocompleteSuggestions = filteredCurrencies.map((currencyName) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.CURRENCY,
-                        text: currencyName,
-                    }));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.TAX_RATE: {
-                    const filteredTaxRates = taxAutocompleteList
-                        .filter((tax) => tax.taxRateName.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(tax.taxRateName.toLowerCase()))
-                        .sort()
-                        .slice(0, 10);
-                    filteredAutocompleteSuggestions = filteredTaxRates.map((tax) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.TAX_RATE,
-                        text: tax.taxRateName,
-                        autocompleteID: tax.taxRateIds.join(','),
-                    }));
-
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.FROM: {
-                    const filteredParticipants = participantsAutocompleteList
-                        .filter(
-                            (participant) => participant.name.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(participant.name.toLowerCase()),
-                        )
-                        .sort()
-                        .slice(0, 10);
-                    filteredAutocompleteSuggestions = filteredParticipants.map((participant) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.FROM,
-                        text: participant.name,
-                        autocompleteID: participant.accountID,
-                    }));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.TO: {
-                    const filteredParticipants = participantsAutocompleteList
-                        .filter(
-                            (participant) => participant.name.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(participant.name.toLowerCase()),
-                        )
-                        .sort()
-                        .slice(0, 10);
-                    filteredAutocompleteSuggestions = filteredParticipants.map((participant) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.TO,
-                        text: participant.name,
-                        autocompleteID: participant.accountID,
-                    }));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.IN: {
-                    const filteredChats = searchOptions.recentReports
-                        .filter((chat) => chat.text?.toLowerCase()?.includes(autocompleteValue.toLowerCase()))
-                        .sort((chatA, chatB) => (chatA > chatB ? 1 : -1))
-                        .slice(0, 10);
-                    filteredAutocompleteSuggestions = filteredChats.map((chat) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.IN,
-                        text: chat.text ?? '',
-                        autocompleteID: chat.reportID,
-                    }));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_ROOT_KEYS.TYPE: {
-                    const filteredTypes = typeAutocompleteList
-                        .filter((type) => type.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(type.toLowerCase()))
-                        .sort();
-                    filteredAutocompleteSuggestions = filteredTypes.map((type) => ({filterKey: CONST.SEARCH.SYNTAX_ROOT_KEYS.TYPE, text: type}));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_ROOT_KEYS.STATUS: {
-                    const filteredStatuses = statusAutocompleteList
-                        .filter((status) => status.includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(status))
-                        .sort()
-                        .slice(0, 10);
-                    filteredAutocompleteSuggestions = filteredStatuses.map((status) => ({filterKey: CONST.SEARCH.SYNTAX_ROOT_KEYS.STATUS, text: status}));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.EXPENSE_TYPE: {
-                    const filteredExpenseTypes = expenseTypes
-                        .filter((expenseType) => expenseType.includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(expenseType))
-                        .sort();
-
-                    filteredAutocompleteSuggestions = filteredExpenseTypes.map((expenseType) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.EXPENSE_TYPE,
-                        text: expenseType,
-                    }));
-                    break;
-                }
-                case CONST.SEARCH.SYNTAX_FILTER_KEYS.CARD_ID: {
-                    const filteredCards = cardAutocompleteList
-                        .filter((card) => card.bank.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(card.bank.toLowerCase()))
-                        .sort()
-                        .slice(0, 10);
-
-                    filteredAutocompleteSuggestions = filteredCards.map((card) => ({
-                        filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.CARD_ID,
-                        text: CardUtils.getCardDescription(card.cardID),
-                        autocompleteID: card.cardID.toString(),
-                    }));
-                    break;
-                }
-                default: {
-                    filteredAutocompleteSuggestions = undefined;
-                }
-            }
-            setAutocompleteSuggestions(filteredAutocompleteSuggestions);
-        },
-        [
-            tagAutocompleteList,
-            recentTagsAutocompleteList,
-            categoryAutocompleteList,
-            recentCategoriesAutocompleteList,
-            currencyAutocompleteList,
-            recentCurrencyAutocompleteList,
-            taxAutocompleteList,
-            participantsAutocompleteList,
-            searchOptions.recentReports,
-            typeAutocompleteList,
-            statusAutocompleteList,
-            expenseTypes,
-            cardAutocompleteList,
-        ],
-    );
-
-    const prevUserQueryRef = useRef<string | null>(null);
     useEffect(() => {
         Report.searchInServer(debouncedInputValue.trim());
     }, [debouncedInputValue]);
 
-    const onSearchChange = useCallback(
+    const sections = [];
+
+    if (contextualReportData && !textInputValue) {
+        const reportQueryValue = contextualReportData.text ?? contextualReportData.alternateText ?? contextualReportData.reportID;
+        sections.push({
+            data: [
+                {
+                    text: `${translate('search.searchIn')} ${contextualReportData.text ?? contextualReportData.alternateText}`,
+                    singleIcon: Expensicons.MagnifyingGlass,
+                    searchQuery: reportQueryValue,
+                    autocompleteID: contextualReportData.reportID,
+                    itemStyle: styles.activeComponentBG,
+                    keyForList: 'contextualSearch',
+                    searchItemType: CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.CONTEXTUAL_SUGGESTION,
+                },
+            ],
+        });
+    }
+
+    const recentSearchesData = sortedRecentSearches?.slice(0, 5).map(({query, timestamp}) => {
+        const searchQueryJSON = SearchQueryUtils.buildSearchQueryJSON(query);
+        return {
+            text: searchQueryJSON ? SearchQueryUtils.buildUserReadableQueryString(searchQueryJSON, personalDetails, cardList, reports, taxRates) : query,
+            singleIcon: Expensicons.History,
+            searchQuery: query,
+            keyForList: timestamp,
+            searchItemType: CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.SEARCH,
+        };
+    });
+
+    if (!textInputValue && recentSearchesData && recentSearchesData.length > 0) {
+        sections.push({title: translate('search.recentSearches'), data: recentSearchesData});
+    }
+
+    const styledRecentReports = recentReports.map((item) => ({...item, pressableStyle: styles.br2, wrapperStyle: [styles.pr3, styles.pl3]}));
+    sections.push({title: translate('search.recentChats'), data: styledRecentReports});
+
+    const searchQueryItem = textInputValue
+        ? {
+              text: textInputValue,
+              singleIcon: Expensicons.MagnifyingGlass,
+              searchQuery: textInputValue,
+              itemStyle: styles.activeComponentBG,
+              keyForList: 'findItem',
+              searchItemType: CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.SEARCH,
+          }
+        : undefined;
+
+    const onSearchQueryChange = useCallback(
         (userQuery: string) => {
-            let newUserQuery = userQuery;
-            if (autocompleteSuggestions && userQuery.endsWith(',')) {
-                newUserQuery = `${userQuery.slice(0, userQuery.length - 1).trim()},`;
+            const prevParsedQuery = parseForAutocomplete(textInputValue);
+
+            let updatedUserQuery = userQuery;
+            // If the prev value was query with autocomplete, and the current query ends with a comma, then we allow to continue autocompleting the next value
+            if (prevParsedQuery?.autocomplete && userQuery.endsWith(',')) {
+                updatedUserQuery = `${userQuery.slice(0, userQuery.length - 1).trim()},`;
             }
-            setTextInputValue(newUserQuery);
-            const autocompleteParsedQuery = parseForAutocomplete(newUserQuery);
-            updateAutocomplete(autocompleteParsedQuery?.autocomplete?.value ?? '', autocompleteParsedQuery?.ranges ?? [], autocompleteParsedQuery?.autocomplete?.key);
+            setTextInputValue(updatedUserQuery);
+            setAutocompleteInputValue(updatedUserQuery);
 
             const updatedSubstitutionsMap = getUpdatedSubstitutionsMap(userQuery, autocompleteSubstitutions);
             setAutocompleteSubstitutions(updatedSubstitutionsMap);
 
-            if (newUserQuery || !isEmpty(prevUserQueryRef.current)) {
+            if (updatedUserQuery) {
                 listRef.current?.updateAndScrollToFocusedIndex(0);
             } else {
                 listRef.current?.updateAndScrollToFocusedIndex(-1);
             }
-
-            // Store the previous newUserQuery
-            prevUserQueryRef.current = newUserQuery;
         },
-        [autocompleteSubstitutions, autocompleteSuggestions, setTextInputValue, updateAutocomplete],
+        [autocompleteSubstitutions, setTextInputValue, textInputValue],
     );
 
     const onSearchSubmit = useCallback(
@@ -369,6 +199,7 @@ function SearchRouter({onRouterClose, shouldHideInputCaret}: SearchRouterProps) 
             Navigation.navigate(ROUTES.SEARCH_CENTRAL_PANE.getRoute({query}));
 
             setTextInputValue('');
+            setAutocompleteInputValue('');
         },
         [autocompleteSubstitutions, onRouterClose, setTextInputValue],
     );
@@ -399,7 +230,7 @@ function SearchRouter({onRouterClose, shouldHideInputCaret}: SearchRouterProps) 
             <SearchRouterInput
                 value={textInputValue}
                 isFullWidth={shouldUseNarrowLayout}
-                updateSearch={onSearchChange}
+                onSearchQueryChange={onSearchQueryChange}
                 onSubmit={() => {
                     onSearchSubmit(textInputValue);
                 }}
@@ -412,13 +243,11 @@ function SearchRouter({onRouterClose, shouldHideInputCaret}: SearchRouterProps) 
                 isSearchingForReports={isSearchingForReports}
             />
             <SearchRouterList
-                textInputValue={textInputValue}
-                updateSearchValue={onSearchChange}
+                textInputValue={autocompleteInputValue}
+                searchQueryItem={searchQueryItem}
+                additionalSections={sections}
                 setTextInputValue={setTextInputValue}
-                reportForContextualSearch={contextualReportData}
-                recentSearches={sortedRecentSearches?.slice(0, 5)}
-                recentReports={recentReports}
-                autocompleteSuggestions={autocompleteSuggestions}
+                onSearchQueryChange={onSearchQueryChange}
                 onSearchSubmit={onSearchSubmit}
                 closeRouter={onRouterClose}
                 onAutocompleteSuggestionClick={onAutocompleteSuggestionClick}
