+++ src/components/Search/SearchRouter/SearchRouterList.tsx
@@ -1,37 +1,44 @@
-import React, {forwardRef, useCallback} from 'react';
+import {Str} from 'expensify-common';
+import React, {forwardRef, useCallback, useMemo} from 'react';
 import type {ForwardedRef} from 'react';
 import {useOnyx} from 'react-native-onyx';
-import type {ValueOf} from 'type-fest';
 import * as Expensicons from '@components/Icon/Expensicons';
 import {usePersonalDetails} from '@components/OnyxProvider';
+import {useOptionsList} from '@components/OptionListContextProvider';
 import type {SearchFilterKey, SearchQueryString} from '@components/Search/types';
 import SelectionList from '@components/SelectionList';
 import SearchQueryListItem from '@components/SelectionList/Search/SearchQueryListItem';
 import type {SearchQueryItem, SearchQueryListItemProps} from '@components/SelectionList/Search/SearchQueryListItem';
 import type {SectionListDataType, SelectionListHandle, UserListItemProps} from '@components/SelectionList/types';
 import UserListItem from '@components/SelectionList/UserListItem';
+import useActiveWorkspace from '@hooks/useActiveWorkspace';
 import useLocalize from '@hooks/useLocalize';
+import usePolicy from '@hooks/usePolicy';
 import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
+import * as CardUtils from '@libs/CardUtils';
 import Navigation from '@libs/Navigation/Navigation';
-import type {SearchOption} from '@libs/OptionsListUtils';
+import * as OptionsListUtils from '@libs/OptionsListUtils';
 import Performance from '@libs/Performance';
 import {getAllTaxRates} from '@libs/PolicyUtils';
 import type {OptionData} from '@libs/ReportUtils';
-import {getQueryWithoutAutocompletedPart} from '@libs/SearchAutocompleteUtils';
+import {
+    getAutocompleteCategories,
+    getAutocompleteRecentCategories,
+    getAutocompleteRecentTags,
+    getAutocompleteTags,
+    getAutocompleteTaxList,
+    getQueryWithoutAutocompletedPart,
+    parseForAutocomplete,
+} from '@libs/SearchAutocompleteUtils';
 import * as SearchQueryUtils from '@libs/SearchQueryUtils';
 import * as ReportUserActions from '@userActions/Report';
 import Timing from '@userActions/Timing';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
-import type Report from '@src/types/onyx/Report';
+import type PersonalDetails from '@src/types/onyx/PersonalDetails';
 import {getSubstitutionMapKey} from './getQueryWithSubstitutions';
-
-type SearchQueryItemData = {
-    query: string;
-    text?: string;
-};
 
 type AutocompleteItemData = {
     filterKey: SearchFilterKey;
@@ -40,29 +47,23 @@ type AutocompleteItemData = {
 };
 
 type SearchRouterListProps = {
-    /** value of TextInput */
+    /** Value of TextInput */
     textInputValue: string;
 
+    /** An optional item to always display on the top of the router list  */
+    searchQueryItem?: SearchQueryItem;
+
+    /** Any extra sections that should be displayed in the router list */
+    additionalSections?: Array<SectionListDataType<OptionData | SearchQueryItem>>;
+
     /** Callback to update text input value along with autocomplete suggestions */
-    updateSearchValue: (newValue: string) => void;
+    onSearchQueryChange: (newValue: string) => void;
 
     /** Callback to update text input value */
     setTextInputValue: (text: string) => void;
 
-    /** Recent searches */
-    recentSearches: Array<SearchQueryItemData & {timestamp: string}> | undefined;
-
-    /** Recent reports */
-    recentReports: OptionData[];
-
-    /** Autocomplete items */
-    autocompleteSuggestions: AutocompleteItemData[] | undefined;
-
     /** Callback to submit query when selecting a list item */
     onSearchSubmit: (query: SearchQueryString) => void;
-
-    /** Context present when opening SearchRouter from a report, invoice or workspace page */
-    reportForContextualSearch?: OptionData;
 
     /** Callback to run when user clicks a suggestion item that contains autocomplete data */
     onAutocompleteSuggestionClick: (autocompleteKey: string, autocompleteID: string) => void;
@@ -132,111 +133,242 @@ function SearchRouterItem(props: UserListItemProps<OptionData> | SearchQueryList
 }
 
 function SearchRouterList(
-    {
-        textInputValue,
-        updateSearchValue,
-        setTextInputValue,
-        reportForContextualSearch,
-        recentSearches,
-        autocompleteSuggestions,
-        recentReports,
-        onSearchSubmit,
-        onAutocompleteSuggestionClick,
-        closeRouter,
-    }: SearchRouterListProps,
+    {textInputValue, searchQueryItem, additionalSections, onSearchQueryChange, setTextInputValue, onSearchSubmit, onAutocompleteSuggestionClick, closeRouter}: SearchRouterListProps,
     ref: ForwardedRef<SelectionListHandle>,
 ) {
     const styles = useThemeStyles();
     const {translate} = useLocalize();
     const {shouldUseNarrowLayout} = useResponsiveLayout();
 
-    const personalDetails = usePersonalDetails();
-    const [reports] = useOnyx(ONYXKEYS.COLLECTION.REPORT);
-    const taxRates = getAllTaxRates();
+    const {activeWorkspaceID} = useActiveWorkspace();
+    const policy = usePolicy(activeWorkspaceID);
+    const [betas] = useOnyx(ONYXKEYS.BETAS);
+
+    const {options, areOptionsInitialized} = useOptionsList();
+    const searchOptions = useMemo(() => {
+        if (!areOptionsInitialized) {
+            return {recentReports: [], personalDetails: [], userToInvite: null, currentUserOption: null, categoryOptions: [], tagOptions: [], taxRatesOptions: []};
+        }
+        return OptionsListUtils.getSearchOptions(options, '', betas ?? []);
+    }, [areOptionsInitialized, betas, options]);
+
+    const typeAutocompleteList = Object.values(CONST.SEARCH.DATA_TYPES);
+    const statusAutocompleteList = Object.values({...CONST.SEARCH.STATUS.TRIP, ...CONST.SEARCH.STATUS.INVOICE, ...CONST.SEARCH.STATUS.CHAT, ...CONST.SEARCH.STATUS.TRIP});
+    const expenseTypes = Object.values(CONST.SEARCH.TRANSACTION_TYPE);
+
     const [cardList = {}] = useOnyx(ONYXKEYS.CARD_LIST);
-    const sections: Array<SectionListDataType<OptionData | SearchQueryItem>> = [];
+    const cardAutocompleteList = Object.values(cardList);
+    const personalDetailsForParticipants = usePersonalDetails();
+    const participantsAutocompleteList = useMemo(
+        () =>
+            Object.values(personalDetailsForParticipants)
+                .filter((details): details is NonNullable<PersonalDetails> => !!(details && details?.login))
+                .map((details) => ({
+                    name: details.displayName ?? Str.removeSMSDomain(details.login ?? ''),
+                    accountID: details?.accountID.toString(),
+                })),
+        [personalDetailsForParticipants],
+    );
+    const taxRates = getAllTaxRates();
+    const taxAutocompleteList = useMemo(() => getAutocompleteTaxList(taxRates, policy), [policy, taxRates]);
+    const [allPolicyCategories] = useOnyx(ONYXKEYS.COLLECTION.POLICY_CATEGORIES);
+    const [allRecentCategories] = useOnyx(ONYXKEYS.COLLECTION.POLICY_RECENTLY_USED_CATEGORIES);
+    const categoryAutocompleteList = useMemo(() => {
+        return getAutocompleteCategories(allPolicyCategories, activeWorkspaceID);
+    }, [activeWorkspaceID, allPolicyCategories]);
+    const recentCategoriesAutocompleteList = useMemo(() => {
+        return getAutocompleteRecentCategories(allRecentCategories, activeWorkspaceID);
+    }, [activeWorkspaceID, allRecentCategories]);
 
-    if (textInputValue) {
-        sections.push({
-            data: [
-                {
-                    text: textInputValue,
-                    singleIcon: Expensicons.MagnifyingGlass,
-                    searchQuery: textInputValue,
-                    itemStyle: styles.activeComponentBG,
-                    keyForList: 'findItem',
-                    searchItemType: CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.SEARCH,
-                },
-            ],
-        });
-    }
+    const [currencyList] = useOnyx(ONYXKEYS.CURRENCY_LIST);
+    const currencyAutocompleteList = Object.keys(currencyList ?? {});
+    const [recentCurrencyAutocompleteList] = useOnyx(ONYXKEYS.RECENTLY_USED_CURRENCIES);
+    const [allPoliciesTags] = useOnyx(ONYXKEYS.COLLECTION.POLICY_TAGS);
+    const [allRecentTags] = useOnyx(ONYXKEYS.COLLECTION.POLICY_RECENTLY_USED_TAGS);
+    const tagAutocompleteList = useMemo(() => {
+        return getAutocompleteTags(allPoliciesTags, activeWorkspaceID);
+    }, [activeWorkspaceID, allPoliciesTags]);
+    const recentTagsAutocompleteList = getAutocompleteRecentTags(allRecentTags, activeWorkspaceID);
 
-    if (reportForContextualSearch && !textInputValue) {
-        const reportQueryValue = reportForContextualSearch.text ?? reportForContextualSearch.alternateText ?? reportForContextualSearch.reportID;
-        let roomType: ValueOf<typeof CONST.SEARCH.DATA_TYPES> = CONST.SEARCH.DATA_TYPES.CHAT;
-        let autocompleteID = reportForContextualSearch.reportID;
-        if (reportForContextualSearch.isInvoiceRoom) {
-            roomType = CONST.SEARCH.DATA_TYPES.INVOICE;
-            const report = reportForContextualSearch as SearchOption<Report>;
-            if (report.item && report.item?.invoiceReceiver && report.item.invoiceReceiver?.type === CONST.REPORT.INVOICE_RECEIVER_TYPE.INDIVIDUAL) {
-                autocompleteID = report.item.invoiceReceiver.accountID.toString();
-            } else {
-                autocompleteID = '';
+    const autocompleteSuggestions = useMemo<AutocompleteItemData[]>(() => {
+        const autocompleteParsedQuery = parseForAutocomplete(textInputValue);
+        const {autocomplete, ranges = []} = autocompleteParsedQuery ?? {};
+        const autocompleteKey = autocomplete?.key;
+        const autocompleteValue = autocomplete?.value ?? '';
+
+        const alreadyAutocompletedKeys = ranges
+            .filter((range) => {
+                return autocompleteKey && range.key === autocompleteKey;
+            })
+            .map((range) => range.value.toLowerCase());
+
+        switch (autocompleteKey) {
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.TAG: {
+                const autocompleteList = autocompleteValue ? tagAutocompleteList : recentTagsAutocompleteList ?? [];
+                const filteredTags = autocompleteList
+                    .filter((tag) => tag.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(tag))
+                    .sort()
+                    .slice(0, 10);
+
+                return filteredTags.map((tagName) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.TAG,
+                    text: tagName,
+                }));
+            }
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.CATEGORY: {
+                const autocompleteList = autocompleteValue ? categoryAutocompleteList : recentCategoriesAutocompleteList;
+                const filteredCategories = autocompleteList
+                    .filter((category) => category.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(category.toLowerCase()))
+                    .sort()
+                    .slice(0, 10);
+
+                return filteredCategories.map((categoryName) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.CATEGORY,
+                    text: categoryName,
+                }));
+            }
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.CURRENCY: {
+                const autocompleteList = autocompleteValue ? currencyAutocompleteList : recentCurrencyAutocompleteList ?? [];
+                const filteredCurrencies = autocompleteList
+                    .filter((currency) => currency.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(currency.toLowerCase()))
+                    .sort()
+                    .slice(0, 10);
+
+                return filteredCurrencies.map((currencyName) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.CURRENCY,
+                    text: currencyName,
+                }));
+            }
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.TAX_RATE: {
+                const filteredTaxRates = taxAutocompleteList
+                    .filter((tax) => tax.taxRateName.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(tax.taxRateName.toLowerCase()))
+                    .sort()
+                    .slice(0, 10);
+
+                return filteredTaxRates.map((tax) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.TAX_RATE,
+                    text: tax.taxRateName,
+                    autocompleteID: tax.taxRateIds.join(','),
+                }));
+            }
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.FROM: {
+                const filteredParticipants = participantsAutocompleteList
+                    .filter((participant) => participant.name.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(participant.name.toLowerCase()))
+                    .sort()
+                    .slice(0, 10);
+
+                return filteredParticipants.map((participant) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.FROM,
+                    text: participant.name,
+                    autocompleteID: participant.accountID,
+                }));
+            }
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.TO: {
+                const filteredParticipants = participantsAutocompleteList
+                    .filter((participant) => participant.name.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(participant.name.toLowerCase()))
+                    .sort()
+                    .slice(0, 10);
+
+                return filteredParticipants.map((participant) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.TO,
+                    text: participant.name,
+                    autocompleteID: participant.accountID,
+                }));
+            }
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.IN: {
+                const filteredChats = searchOptions.recentReports
+                    .filter((chat) => chat.text?.toLowerCase()?.includes(autocompleteValue.toLowerCase()))
+                    .sort((chatA, chatB) => (chatA > chatB ? 1 : -1))
+                    .slice(0, 10);
+
+                return filteredChats.map((chat) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.IN,
+                    text: chat.text ?? '',
+                    autocompleteID: chat.reportID,
+                }));
+            }
+            case CONST.SEARCH.SYNTAX_ROOT_KEYS.TYPE: {
+                const filteredTypes = typeAutocompleteList
+                    .filter((type) => type.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(type.toLowerCase()))
+                    .sort();
+
+                return filteredTypes.map((type) => ({filterKey: CONST.SEARCH.SYNTAX_ROOT_KEYS.TYPE, text: type}));
+            }
+            case CONST.SEARCH.SYNTAX_ROOT_KEYS.STATUS: {
+                const filteredStatuses = statusAutocompleteList
+                    .filter((status) => status.includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(status))
+                    .sort()
+                    .slice(0, 10);
+
+                return filteredStatuses.map((status) => ({filterKey: CONST.SEARCH.SYNTAX_ROOT_KEYS.STATUS, text: status}));
+            }
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.EXPENSE_TYPE: {
+                const filteredExpenseTypes = expenseTypes
+                    .filter((expenseType) => expenseType.includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(expenseType))
+                    .sort();
+
+                return filteredExpenseTypes.map((expenseType) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.EXPENSE_TYPE,
+                    text: expenseType,
+                }));
+            }
+            case CONST.SEARCH.SYNTAX_FILTER_KEYS.CARD_ID: {
+                const filteredCards = cardAutocompleteList
+                    .filter((card) => card.bank.toLowerCase().includes(autocompleteValue.toLowerCase()) && !alreadyAutocompletedKeys.includes(card.bank.toLowerCase()))
+                    .sort()
+                    .slice(0, 10);
+
+                return filteredCards.map((card) => ({
+                    filterKey: CONST.SEARCH.SYNTAX_FILTER_KEYS.CARD_ID,
+                    text: CardUtils.getCardDescription(card.cardID),
+                    autocompleteID: card.cardID.toString(),
+                }));
+            }
+            default: {
+                return [];
             }
         }
-        if (reportForContextualSearch.isPolicyExpenseChat) {
-            roomType = CONST.SEARCH.DATA_TYPES.EXPENSE;
-            autocompleteID = reportForContextualSearch.policyID ?? '';
-        }
-        sections.push({
-            data: [
-                {
-                    text: `${translate('search.searchIn')} ${reportForContextualSearch.text ?? reportForContextualSearch.alternateText}`,
-                    singleIcon: Expensicons.MagnifyingGlass,
-                    searchQuery: reportQueryValue,
-                    autocompleteID,
-                    itemStyle: styles.activeComponentBG,
-                    keyForList: 'contextualSearch',
-                    searchItemType: CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.CONTEXTUAL_SUGGESTION,
-                    roomType,
-                    policyID: reportForContextualSearch.policyID,
-                },
-            ],
-        });
+    }, [
+        textInputValue,
+        tagAutocompleteList,
+        recentTagsAutocompleteList,
+        categoryAutocompleteList,
+        recentCategoriesAutocompleteList,
+        currencyAutocompleteList,
+        recentCurrencyAutocompleteList,
+        taxAutocompleteList,
+        participantsAutocompleteList,
+        searchOptions.recentReports,
+        typeAutocompleteList,
+        statusAutocompleteList,
+        expenseTypes,
+        cardAutocompleteList,
+    ]);
+
+    const sections: Array<SectionListDataType<OptionData | SearchQueryItem>> = [];
+
+    if (searchQueryItem) {
+        sections.push({data: [searchQueryItem]});
     }
 
-    const autocompleteData = autocompleteSuggestions?.map(({filterKey, text, autocompleteID}) => {
-        return {
-            text: getSubstitutionMapKey(filterKey, text),
-            singleIcon: Expensicons.MagnifyingGlass,
-            searchQuery: text,
-            autocompleteID,
-            keyForList: autocompleteID ?? text, // in case we have a unique identifier then use it because text might not be unique
-            searchItemType: CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.AUTOCOMPLETE_SUGGESTION,
-        };
-    });
+    if (autocompleteSuggestions.length > 0) {
+        const autocompleteData = autocompleteSuggestions.map(({filterKey, text, autocompleteID}) => {
+            return {
+                text: getSubstitutionMapKey(filterKey, text),
+                singleIcon: Expensicons.MagnifyingGlass,
+                searchQuery: text,
+                autocompleteID,
+                keyForList: autocompleteID ?? text, // in case we have a unique identifier then use it because text might not be unique
+                searchItemType: CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.AUTOCOMPLETE_SUGGESTION,
+            };
+        });
 
-    if (autocompleteData && autocompleteData.length > 0) {
         sections.push({title: translate('search.suggestions'), data: autocompleteData});
     }
 
-    const recentSearchesData = recentSearches?.map(({query, timestamp}) => {
-        const searchQueryJSON = SearchQueryUtils.buildSearchQueryJSON(query);
-        return {
-            text: searchQueryJSON ? SearchQueryUtils.buildUserReadableQueryString(searchQueryJSON, personalDetails, cardList, reports, taxRates) : query,
-            singleIcon: Expensicons.History,
-            searchQuery: query,
-            keyForList: timestamp,
-            searchItemType: CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.SEARCH,
-        };
-    });
-
-    if (!textInputValue && recentSearchesData && recentSearchesData.length > 0) {
-        sections.push({title: translate('search.recentSearches'), data: recentSearchesData});
+    if (additionalSections) {
+        sections.push(...additionalSections);
     }
-
-    const styledRecentReports = recentReports.map((item) => ({...item, pressableStyle: styles.br2, wrapperStyle: [styles.pr3, styles.pl3]}));
-    sections.push({title: translate('search.recentChats'), data: styledRecentReports});
 
     const onSelectRow = useCallback(
         (item: OptionData | SearchQueryItem) => {
@@ -246,7 +378,7 @@ function SearchRouterList(
                 }
                 if (item.searchItemType === CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.CONTEXTUAL_SUGGESTION) {
                     const searchQuery = getContextualSearchQuery(item);
-                    updateSearchValue(`${searchQuery} `);
+                    onSearchQueryChange(`${searchQuery} `);
 
                     if (item.roomType === CONST.SEARCH.DATA_TYPES.INVOICE && item.autocompleteID) {
                         const autocompleteKey = `${CONST.SEARCH.SYNTAX_FILTER_KEYS.TO}:${item.searchQuery}`;
@@ -260,7 +392,7 @@ function SearchRouterList(
                 }
                 if (item.searchItemType === CONST.SEARCH.SEARCH_ROUTER_ITEM_TYPE.AUTOCOMPLETE_SUGGESTION && textInputValue) {
                     const trimmedUserSearchQuery = getQueryWithoutAutocompletedPart(textInputValue);
-                    updateSearchValue(`${trimmedUserSearchQuery}${SearchQueryUtils.sanitizeSearchValue(item.searchQuery)} `);
+                    onSearchQueryChange(`${trimmedUserSearchQuery}${SearchQueryUtils.sanitizeSearchValue(item.searchQuery)} `);
 
                     if (item.autocompleteID && item.text) {
                         onAutocompleteSuggestionClick(item.text, item.autocompleteID);
@@ -279,7 +411,7 @@ function SearchRouterList(
                 ReportUserActions.navigateToAndOpenReport(item.login ? [item.login] : [], false);
             }
         },
-        [closeRouter, textInputValue, onSearchSubmit, updateSearchValue, onAutocompleteSuggestionClick],
+        [closeRouter, textInputValue, onSearchSubmit, onSearchQueryChange, onAutocompleteSuggestionClick],
     );
 
     const onArrowFocus = useCallback(
@@ -289,6 +421,7 @@ function SearchRouterList(
             }
 
             const trimmedUserSearchQuery = getQueryWithoutAutocompletedPart(textInputValue);
+
             setTextInputValue(`${trimmedUserSearchQuery}${SearchQueryUtils.sanitizeSearchValue(focusedItem.searchQuery)} `);
 
             if (focusedItem.autocompleteID && focusedItem.text) {
@@ -319,4 +452,3 @@ function SearchRouterList(
 
 export default forwardRef(SearchRouterList);
 export {SearchRouterItem};
-export type {AutocompleteItemData};
